hw/phb4: Use standard MIN/MAX macro definitions
The max() macro definition incorrectly returns the minimum value.  The
max() macro is used to ensure that PERST has been asserted for 250ms and
that we wait 100ms seconds for the ETU logic in the CRESET_START PHB4
PCI slot state.  However, by returning the minimum value there is no
guarantee that either of these requirements are met.

Correct macro definitions for MIN and MAX are already provided in
skiboot.h.  Remove the redundant/incorrect versions here and switch to
using the standard ones.

Fixes: 70edcbb4b39d ("hw/phb4: Skip FRESET PERST when coming from
CRESET")
Signed-off-by: Jordan Niethe <jniethe5@gmail.com>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>