npu2: Invalidate entire TCE cache if many entries requested
[ Upstream commit 2a0455ba0f7784b2d7e9e3915fd30f815afd2ae1 ]

Turned out invalidating entries in NPU TCE cache is so slow that it
becomes visible when running a 30+GB guest with GPU+NVlink2 passed
through; a 100GB guest takes about 20s to map all 100GB.

This falls through to the entire cache invalidation if more than 128
TCEs were requested to invalidate, this reduces 20s from the abobe to
less than 1s. The KVM change [1] is required to see this difference.

The threshold of 128 is chosen in attempt not to affect performance much
as it is not clear how expensive it is to populate the TCE cache again;
all we know for sure is that mapping the guest produces invalidation
requests of 512 TCEs each.

Note TCE cache invalidation in PHB4 is faster and does not require
the same workaround.

[1] KVM: PPC: vfio/spapr_tce: Split out TCE invalidation from TCE updates
https://patchwork.ozlabs.org/patch/1149003/
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Reviewed-by: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>