npu3: Delay enablement of DL parity checking
Currently, we turn on TX and PRI parity checking of the DL during NPU
initialization, while RX parity checking is not enabled until after link
training.

This behavior was prescribed for npu2, but on npu3 systems the logic has
changed such that we're getting early parity error checkstops.

To fix, only set the TX and PRI enable bits after training, consistent
with RX.

Signed-off-by: Reza Arbab <arbab@linux.ibm.com>