npu2-opencapi: Fix integer promotion bug in LPC allocation
[ Upstream commit e85e2e2b8b0a4dd96e10343df19df1a161cd6aac ]

If you try to allocate an amount of LPC memory that's not a power of 2,
we round the value up to the nearest power of 2.

By the magic of C, "1 << n" gets treated as an int, even if you're
assigning it to a uint64_t.

Change 1 to 1ULL to fix this.

(C, it's great.)

Reported-by: Alastair D'Silva <alistair@d-silva.org>
Cc: skiboot-stable@lists.ozlabs.org
Signed-off-by: Andrew Donnellan <ajd@linux.ibm.com>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>