platforms/swift: Remove spurious error message
npu3_chip_possible_gpus() works by dividing the number of NVLink-mode
bricks by the number of bricks connecting a single GPU. In a system with
no GPUs, the latter value is unknown, so the function returns zero and
we trip a somewhat misleading error message.

The code afterward is safe to execute in any case, so there's no need to
return either. Remove the check entirely.

Signed-off-by: Reza Arbab <arbab@linux.ibm.com>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>