platforms/qemu: Add slot table
Add a slot table to our QEMU platform file to allow testing slot
functions and so that we have the basic infrastructure in place. The
slot names are the bus names that Qemu assigns to the subordiate bus of the
root port (i.e. where actual devices should go).

Devices can be added under a specific phb by specifying that bus. E.g.

        -drive file=./disk1.img,format=raw,if=none,id=nvme0 \
        -device nvme,drive=nvme0,bus=pcie.1

Places an NVMe device under the root port of PHB0001.

The six slots here will cover each of the six PHBs that exist in the
default P9 model. The P8 model has three (since that's what a Venice
has), but the extra slots shouldn't cause any problems.

Reviewed-by: Cédric Le Goater <clg@kaod.org>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>