test/mambo: Skip restoring the old ulimit
Reducing the core size limit is always permitted, but increasing it may
not. To work around problems with some versions of mambo segfaulting we
set the ulimit to zero when running the tests and try to restore the old
ulimit afterwards.

On systems where this isn't permitted (fedora?) this results in a pile
of of spurious errors from ulimit while running "make check. e.g.

   ulimit: core file size: cannot modify limit: Operation not permitted

The ulimit reduction only applies to the currently running shell (i.e
the one running the script) which is probably not going to crash outside
of mambo. Remove the ulimit restoration to squash the errors so the
output of make check is cleaner.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>