opal-ci: Remove clang builds on 18.04
commit 3734172b0f40 ("build: fix non-constant build asserts") added the
following check in asm/head.S:

.if reset_patch_end - reset_patch_start > 0x100
       .error "Reset patch overflow"
.endif

Some older versions of clang such as the one that ships with ubuntu
18.04 (vlang 6.0.0) don't consider symbol addresses as constant which
causes the following error:

 asm/head.S:578:5: error: expected absolute expression
 .if reset_patch_end - reset_patch_start > 0x100
     ^
 asm/head.S:579:2: error: Reset patch overflow
  .error "Reset patch overflow"

This appears to be a bug with the clang assembler that is fixed with
versions of clang. Clang support in skiboot is best-effort so removing
the clang builds from the Ubuntu 18.04 CI script is preferable to
working around the limitations of the older versions of clang.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>