core/init: Add ibm,processor-storage-keys property to CPU DT node
LoPAPR says:

    “ibm,processor-storage-keys”

    property name indicating the number of virtual storage keys supported
    by the processor described by this node.

    prop-encoded-array: Consists of two cells encoded as with encode-int.
    The first cell represents the number of virtual storage keys supported
    for data accesses while the second cell represents the number of
    virtual storage keys supported for instruction accesses. The cell value
    of zero indicates that no storage keys are supported for the access
    type.

pHyp provides the property above but there's a bug in P8 firmware
where the second cell is zero even though POWER8 supports instruction
access keys. This bug will be fixed for P9.

While this is a PAPR property, it's useful to have it in powernv as
well so that Linux has a uniform way of checking for the feature
regardless of the platform it's running on.

Tested on QEMU POWER8 powernv model, Mambo P9 and POWER8 Firenze
machine.

Signed-off-by: Thiago Jung Bauermann <bauerman@linux.vnet.ibm.com>
[mpe: Rebase, add comment explaining why we hard code 32]
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>