doc: Fix code blocks in v6.3.3 release notes
Sphinx seems to choke if there's an additional indentation in a ::
block.

e.g.:

::
   one
   two
    three
   four

It'll complain about the indentation changing at three. A
".. code-block:: text" block doesn't seem to have this problem so use
that instead. Also note that you need a blank line between the
code-block and the start of the actual code block.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>