npu2-opencapi: Allow platforms to identify physical slots
[ Upstream commit 9de4f2284c54433f7f4ff3dc3d13a39c657e2c19 ]

This patch lets each platform define the name of the opencapi
slots. It makes it easier to identify which physical card is
generating errors or messages in the linux or skiboot log files.

The patch provides slot names for mihawk and witherspoon. If the
platform doesn't define any, then we default to 'OPENCAPI-xxxx'

There are various ways to find out about the slot names:
  skiboot log
  lspci command (if the PCI hotplug driver pnv-php is loaded)
  lshw
  checking the device tree
  and probably others....

Signed-off-by: Frederic Barrat <fbarrat@linux.ibm.com>
Reviewed-by: Andrew Donnellan <ajd@linux.ibm.com>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>