ibm-fsp/lxvpd: check for upstream port on slot labeling
Certain FSP configurations include PCIe switches that can have LXVPD
slot map entries using the same switch-id and dev-id, even if they are
referring to different upstream and downstream ports of the same link.

The slot matching function (lxvpd_get_slot()) will match the first
occurence, that can be the upstream port with, and ignore the downstream
port.

The main symptom for the above is an incorrect label for those slots,
but I believe other slot attributes could be incorrect as well (as we
are associating a slot with an upstream port).

This patch picks-up an existing "upstream port" attribute from the 1005
version of the LXVPD slot map to prevent matching upstream ports on
the slot matching function.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>
[oliver: 80cols compliance]
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>